chore: feature list data shifted to common
